+++ pcs_test/tier0/lib/commands/test_booth.py
@@ -3803,3 +3803,228 @@ class GetStatusWarnings(TestCase, FixtureMixin):
                 )
             ]
         )
+
+
+class CrmTicketOperationTest:
+    CIB_STATUS = """
+        <status>
+            <tickets>
+                <ticket_state id="T1" granted="false" booth-cfg-name="booth"/>
+                <ticket_state id="T2" granted="false" booth-cfg-name="booth"/>
+                <ticket_state id="T3" granted="false" booth-cfg-name="custom_booth"/>
+                <ticket_state id="T-self-managed" granted="false"/>
+            </tickets>
+        </status>
+    """
+
+    def setUp(self):
+        self.env_assist, self.config = get_env_tools(self)
+        self.config.runner.cib.load(status=self.CIB_STATUS)
+
+    def _call_cmd(self, ticket_name):
+        raise NotImplementedError()
+
+    def fixture_crm_call(self, ticket_name):
+        raise NotImplementedError()
+
+    def fixture_reports(self, ticket_name):
+        raise NotImplementedError()
+
+    def test_default_instance(self):
+        ticket = "T2"
+        self.fixture_crm_call(ticket)
+        self._call_cmd(ticket)
+        self.env_assist.assert_reports(self.fixture_reports(ticket))
+
+    def test_custom_instance(self):
+        ticket = "T3"
+        self.fixture_crm_call(ticket)
+
+        self._call_cmd(ticket)
+        self.env_assist.assert_reports(self.fixture_reports(ticket))
+
+    def test_missing_ticket(self):
+        ticket = "T4"
+        self.env_assist.assert_raise_library_error(
+            lambda: self._call_cmd(ticket)
+        )
+        self.env_assist.assert_reports(
+            [
+                fixture.error(
+                    reports.codes.BOOTH_TICKET_NOT_IN_CIB, ticket_name=ticket
+                )
+            ]
+        )
+
+
+class TicketCleanup(CrmTicketOperationTest, TestCase):
+    def setUp(self):
+        self.env_assist, self.config = get_env_tools(self)
+        self.config.runner.cib.load(status=self.CIB_STATUS)
+
+    def fixture_crm_call(self, ticket_name):
+        self.config.runner.pcmk.ticket_standby(ticket_name)
+        self.config.runner.pcmk.ticket_cleanup(ticket_name)
+
+    def fixture_reports(self, ticket_name):
+        return [
+            fixture.info(
+                reports.codes.BOOTH_TICKET_CHANGING_STATE,
+                ticket_name=ticket_name,
+                state="standby",
+            ),
+            fixture.info(
+                reports.codes.BOOTH_TICKET_CLEANUP, ticket_name=ticket_name
+            ),
+        ]
+
+    def _call_cmd(self, ticket_name):
+        commands.ticket_cleanup(self.env_assist.get_env(), ticket_name)
+
+    def test_standby_fails(self):
+        self.config.runner.pcmk.ticket_standby(
+            "T2", returncode=1, stderr="some error"
+        )
+
+        self.env_assist.assert_raise_library_error(
+            lambda: commands.ticket_cleanup(self.env_assist.get_env(), "T2")
+        )
+        self.env_assist.assert_reports(
+            [
+                fixture.info(
+                    reports.codes.BOOTH_TICKET_CHANGING_STATE,
+                    ticket_name="T2",
+                    state="standby",
+                ),
+                fixture.error(
+                    reports.codes.BOOTH_TICKET_OPERATION_FAILED,
+                    operation="standby",
+                    reason="some error",
+                    site_ip=None,
+                    ticket_name="T2",
+                ),
+            ]
+        )
+
+    def test_cleanup_fails(self):
+        self.config.runner.pcmk.ticket_standby("T2")
+        self.config.runner.pcmk.ticket_cleanup(
+            "T2", returncode=1, stderr="some error"
+        )
+
+        self.env_assist.assert_raise_library_error(
+            lambda: commands.ticket_cleanup(self.env_assist.get_env(), "T2")
+        )
+        self.env_assist.assert_reports(
+            [
+                fixture.info(
+                    reports.codes.BOOTH_TICKET_CHANGING_STATE,
+                    ticket_name="T2",
+                    state="standby",
+                ),
+                fixture.info(
+                    reports.codes.BOOTH_TICKET_CLEANUP, ticket_name="T2"
+                ),
+                fixture.error(
+                    reports.codes.BOOTH_TICKET_OPERATION_FAILED,
+                    operation="cleanup",
+                    reason="some error",
+                    site_ip=None,
+                    ticket_name="T2",
+                ),
+            ]
+        )
+
+
+class TicketUnstandby(CrmTicketOperationTest, TestCase):
+    def setUp(self):
+        self.env_assist, self.config = get_env_tools(self)
+        self.config.runner.cib.load(status=self.CIB_STATUS)
+
+    def fixture_crm_call(self, ticket_name):
+        self.config.runner.pcmk.ticket_unstandby(ticket_name)
+
+    def fixture_reports(self, ticket_name):
+        return [
+            fixture.info(
+                reports.codes.BOOTH_TICKET_CHANGING_STATE,
+                ticket_name=ticket_name,
+                state="active",
+            )
+        ]
+
+    def _call_cmd(self, ticket_name):
+        commands.ticket_unstandby(self.env_assist.get_env(), ticket_name)
+
+    def test_unstandby_fails(self):
+        ticket = "T2"
+        self.config.runner.pcmk.ticket_unstandby(
+            ticket, returncode=1, stderr="some error"
+        )
+
+        self.env_assist.assert_raise_library_error(
+            lambda: commands.ticket_unstandby(self.env_assist.get_env(), ticket)
+        )
+        self.env_assist.assert_reports(
+            [
+                fixture.info(
+                    reports.codes.BOOTH_TICKET_CHANGING_STATE,
+                    ticket_name=ticket,
+                    state="active",
+                ),
+                fixture.error(
+                    reports.codes.BOOTH_TICKET_OPERATION_FAILED,
+                    operation="unstandby",
+                    reason="some error",
+                    site_ip=None,
+                    ticket_name=ticket,
+                ),
+            ]
+        )
+
+
+class TicketStandby(CrmTicketOperationTest, TestCase):
+    def setUp(self):
+        self.env_assist, self.config = get_env_tools(self)
+        self.config.runner.cib.load(status=self.CIB_STATUS)
+
+    def fixture_crm_call(self, ticket_name):
+        self.config.runner.pcmk.ticket_standby(ticket_name)
+
+    def fixture_reports(self, ticket_name):
+        return [
+            fixture.info(
+                reports.codes.BOOTH_TICKET_CHANGING_STATE,
+                ticket_name=ticket_name,
+                state="standby",
+            )
+        ]
+
+    def _call_cmd(self, ticket_name):
+        commands.ticket_standby(self.env_assist.get_env(), ticket_name)
+
+    def test_standby_fails(self):
+        ticket = "T2"
+        self.config.runner.pcmk.ticket_standby(
+            ticket, returncode=1, stderr="some error"
+        )
+
+        self.env_assist.assert_raise_library_error(
+            lambda: commands.ticket_standby(self.env_assist.get_env(), ticket)
+        )
+        self.env_assist.assert_reports(
+            [
+                fixture.info(
+                    reports.codes.BOOTH_TICKET_CHANGING_STATE,
+                    ticket_name=ticket,
+                    state="standby",
+                ),
+                fixture.error(
+                    reports.codes.BOOTH_TICKET_OPERATION_FAILED,
+                    operation="standby",
+                    reason="some error",
+                    site_ip=None,
+                    ticket_name=ticket,
+                ),
+            ]
+        )